refactor: Action 및 Coin 모델의 속성 이름 변경 및 Member 모델에 잔고 감소 메서드 추가

+++ models/dto/decision.py
@@ -4,7 +4,7 @@ class Decision:
         self.down_chance = decision["down_chance"]
         self.details = decision["details"]
         
-    def set_current_price(self, current_price: float):
+    def set_current_price(self, current_price: int):
         self.current_price = current_price
         
     def set_market(self, market: str):
